+++ src/dictionary/DefaultMod.ts
@@ -5,8 +5,9 @@ import TipValue, { TipValueInitOption } from "../lib/TipValue"
 import { LocalValue, LocalValueInitOption, createLocalValue } from "../lib/AttrsValue"
 import InterfaceValue from "../lib/InterfaceValue"
 import { ArrayMapValueType } from "../lib/ArrayMap"
-import InterfaceLayoutValue, { InterfaceLayoutValueInitOption } from "../lib/InterfaceLayoutValue"
 import { renderType } from "../type"
+import { GridValue, buildGridValue } from "../lib/GridParse"
+import WidthValue, { WidthValueInitOption } from "../lib/WidthValue"
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type reactiveFunction = (...args: any[]) => boolean
@@ -16,9 +17,10 @@ export interface DefaultModInitOption extends SimpleDataInitOption {
   $redirect?: string // 快捷格式化目标，内存指针指向对应的mod
   prop?: string
   name?: string
-  layout?: InterfaceLayoutValueInitOption
-  local?: LocalValueInitOption
   tip?: TipValueInitOption
+  grid?: GridValue
+  width?: WidthValueInitOption
+  local?: LocalValueInitOption
   reactives?: Record<string, undefined | reactiveFunction>
   renders?: Record<string, undefined | renderType>
   observe?: observeType
@@ -29,8 +31,9 @@ class DefaultMod extends SimpleData implements ArrayMapValueType {
   static $formatConfig = { name: 'DefaultMod', level: 40, recommend: true }
   $prop: string
   $name: InterfaceValue<string>
-  $layout?: InterfaceLayoutValue
-  tip?: TipValue
+  $tip?: TipValue
+  $grid?: GridValue
+  $width?: WidthValue
   $local?: LocalValue
   $reactives?: Record<string, undefined | reactiveFunction>
   $renders?: Record<string, undefined | renderType>
@@ -42,13 +45,10 @@ class DefaultMod extends SimpleData implements ArrayMapValueType {
     this.$prop = initOption.prop || (parent ? parent.$prop : '')
     this.$name = (initOption.name !== undefined || !parent) ? new InterfaceValue(initOption.name) : parent.$getInterfaceData('name')
     if (initOption.tip !== undefined) {
-      this.tip = new TipValue(initOption.tip)
+      this.$tip = new TipValue(initOption.tip)
     }
-    if (initOption.layout) {
-      this.$layout = new InterfaceLayoutValue(initOption.layout)
-    } else if (parent && parent.$layout) {
-      this.$layout = parent.$layout
-    }
+    this.$grid = buildGridValue(initOption.grid)
+    this.$width = new WidthValue(initOption.width)
     this.$local = createLocalValue(initOption.local)
     this.$reactives = initOption.reactives
     this.$renders = initOption.renders